[ipc] move connection related configuration outside of validator config (#57)
